AB#36 Game working
